Rollback chromedriver version used in Docker to 2.38
Our Docker image uses Chromium v57, which is not supported by
chromedriver versions after 2.38. See
https://bugs.chromium.org/p/chromedriver/issues/detail?id=2445

The reason that the version of chromium is so old is that our base
image ruby:2.3.6 (the standard image for Ruby 2.3.6 on DockerHub) is
built on Debian 8 (jessie), which was made obsolete over a year ago.

The best solution is probably to upgrade the application to Ruby 2.3.7
(if not later), so we can use the official ruby:2.3.7 image, which is
kept up-to-date (it has Chromium 68).

Alternatively, we could try to install Chromium from a more up-to-date
repository, but this is probably not worth the trouble.