The value of cweek did not manage well the last week of the year
grouping_key_for method when params[:group] == "week" always returned "#{date.cweek}/#{date.year}" but that not always is true.
When a date belongs to the first_week/last_week of the year we have exceptions for example "31/12/2018".  His cweek is equal to "1", old code returned "1/2018" value but correct result would be "1/2019".