Replace `attribute_changed?` in `after` callbacks
This method is deprecated in Rails 5.1 because its behavior will be
different in `before` and `after` callbacks.

We're replacing the deprecated `attribute_changed?`, `attribute_was` and
`changed_attributes` methods with `saved_change_to_attribute?`,
`attribute_before_last_save` and `saved_changes` during `after_save`
callbacks.

Note we can't replace this method in the `change_log` callback in
investments because it uses a translatable field, and globalize doesn't
work properly with the new behavior of dirty attributes in Rails 5.1.

https://github.com/rails/rails/pull/32835/