Use the same view code for finished budgets
The differences between the custom code and the code in CONSUL didn't
make much sense anymore: the path can be passed a budget instead of an
ID (so it uses a slug) and there's no reason to hide the link to the
executions for the 2018 budget anymore. Furthermore, removing the
`status: 1` parameter makes the link consistent with the rest of the
application, since it was the only link to budget executions using that
parameter.