Order settings to follow our traditional order
Settings are displayed in the administration in the order they are
created, and reordering them alphabetically in the code caused them to
be displayed in a different order.

Since we don't want to confuse our users by randomly changing the order,
I've grouped the settings by namespaces and ordered each namespace the
way it used to be.

Note dev seeds don't need to be changed, since they are executed after
default settings have been created.