Reduce severity of HasManyOrHasOneDependent rule
This is actually a hack. We want Hound to warn us about this rule;
however, we don't want to be notified about our many existing offenses.

Ideally we would add the `dependent` option to all existing models.
However, this is extra tricky because adding this option would change
existing behavior.