Fix flaky unselecting an investment spec
In this spec, we were doing to request at the same time; one to unselect
an investment and another one to filter the investments. If the second
request finished before the first one, the test failed.

Adding an expectation to check the first request has finished before the
second one starts solves the problem.